% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA1.R
\name{DATA1}
\alias{DATA1}
\title{Data generating process: Continuous Case Treatment}
\usage{
DATA1(N, p, theta)
}
\arguments{
\item{N}{The sample size}

\item{p}{The number of associated covariates}

\item{theta}{The true value of the targeted parameter}
}
\description{
Simulates data set that can used in DML and DSL analysis when the treatment variable is considered to be continuous.
}
\examples{
DataSet <- DATA1(N=100, p=5, theta=0.5)
}
