% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA2.R
\name{DATA2}
\alias{DATA2}
\title{Data generating process: Binary Case Treatment}
\usage{
DATA2(N, p, theta)
}
\arguments{
\item{N}{The sample size}

\item{p}{The number of associated covariates}

\item{theta}{The true value of the targeted parameter}
}
\description{
Simulates data set that can used in DML and DSL analysis when the treatment variable is considered to be binary.
}
\examples{
DataSet <- DATA2(N=100, p=5, theta=0.5)
}
