% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DML1.R
\name{DML1}
\alias{DML1}
\title{The Double Machine-Learning Method (DML): Continuous Case Treatment}
\usage{
DML1(data, y.col.n, d.col.n)
}
\arguments{
\item{data}{The data set. Treatment variable D must be continuous, and the number of variables must be greater than 3.}

\item{y.col.n}{The column number in the data set that represents the response.}

\item{d.col.n}{The column number in the data set that represents the treatment variable.}
}
\description{
Estimates the treatment effect using the DML method when the treatment variable is considered to be continuous.
}
\examples{
DataSet <- DATA1(N=100, p=5, theta=0.5)
DML1(data=DataSet, y.col.n=1, d.col.n=2)
}
