% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DML2.R
\name{DML2}
\alias{DML2}
\title{The Double Machine-Learning Method (DML): Binary Case Treatment}
\usage{
DML2(data, y.col.n, d.col.n)
}
\arguments{
\item{data}{The data set. Treatment variable D must be binary, and the number of variables must be greater than 3.}

\item{y.col.n}{The column number in the data set that represents the response.}

\item{d.col.n}{The column number in the data set that represents the treatment variable.}
}
\description{
Estimates the treatment effect using the DML method when the treatment variable is considered to be binary.
}
\examples{
DataSet <- DATA2(N=100, p=5, theta=0.5)
DML1(data=DataSet, y.col.n=1, d.col.n=2)
}
