% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DSL2.R
\name{DSL2}
\alias{DSL2}
\title{The Double Super Learner Method (DSL): Binary Case Treatment}
\usage{
DSL2(data, y.col.n, d.col.n, SL.library, cv)
}
\arguments{
\item{data}{The data set. Treatment variable D must be binary, and the number of variables must be greater than 3. Variables must be numeric with no missing values.}

\item{y.col.n}{The column number in the data set that represents the response.}

\item{d.col.n}{The column number in the data set that represents the treatment variable.}

\item{SL.library}{The candidate machine-learning algorithms. Must be uploaded first. See example.}

\item{cv}{The number of cross-validation folds followed by the letter L. See example.}
}
\description{
Estimates the treatment effect using the DSL method when the treatment variable is considered to be binary.
}
\examples{
DataSet <- DATA2(N=100, p=5, theta=0.5)
SL.library <- c("SL.biglasso", "SL.glm", "SL.kernelKnn", "SL.ranger", "SL.xgboost")
DSL2(data=DataSet, y.col.n=1, d.col.n=2, SL.library=SL.library, cv=3L)
}
