% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Example2}
\alias{Example2}
\title{Communities and Crime Dataset}
\format{
A dataset for high-dimensional example with 123 rows and 127 variable to perform the DML and DSL methods:
\describe{
\item{x.state}{US state by number - not counted as predictive above, but if considered, should be consided nominal}
\item{x.county}{numeric code for county - not predictive, and many missing values}
\item{x.community}{numeric code for community - not predictive and many missing values}
\item{x.fold}{fold number for non-random 10 fold cross validation, potentially useful for debugging, paired tests - not predictive}
\item{x.population}{population for community}
\item{x.householdsize}{mean people per household}
\item{x.racepctblack}{percentage of population that is african american}
\item{x.racePctWhite}{percentage of population that is caucasian}
\item{x.racePctAsian}{percentage of population that is of asian heritage}
\item{x.racePctHisp}{percentage of population that is of hispanic heritage}
\item{x.agePct12t21}{percentage of population that is 12-21 in age}
\item{x.agePct12t29}{percentage of population that is 12-29 in age}
\item{x.agePct16t24}{percentage of population that is 16-24 in age}
\item{x.agePct65up}{percentage of population that is 65 and over in age}
\item{x.numbUrban}{number of people living in areas classified as urban}
\item{x.pctUrban}{percentage of people living in areas classified as urban}
\item{x.medIncome}{median household income}
\item{x.pctWWage}{percentage of households with wage or salary income in 1989}
\item{x.pctWFarmSelf}{percentage of households with farm or self employment income in 1989}
\item{x.pctWInvInc}{percentage of households with investment / rent income in 1989}
\item{x.pctWSocSec}{percentage of households with social security income in 1989}
\item{x.pctWPubAsst}{percentage of households with public assistance income in 1989}
\item{x.pctWRetire}{percentage of households with retirement income in 1989}
\item{x.medFamInc}{median family income (differs from household income for non-family households)}
\item{x.perCapInc}{per capita income}
\item{x.whitePerCap}{per capita income for caucasians}
\item{x.blackPerCap}{per capita income for african americans}
\item{x.indianPerCap}{per capita income for native americans}
\item{x.AsianPerCap}{per capita income for people with asian heritage}
\item{x.OtherPerCap}{per capita income for people with 'other' heritage}
\item{x.HispPerCap}{per capita income for people with hispanic heritage}
\item{x.NumUnderPov}{number of people under the poverty level}
\item{x.PctPopUnderPov}{percentage of people under the poverty level}
\item{x.PctLess9thGrade}{percentage of people 25 and over with less than a 9th grade education}
\item{x.PctNotHSGrad}{percentage of people 25 and over that are not high school graduates}
\item{x.PctBSorMore}{percentage of people 25 and over with a bachelors degree or higher education}
\item{d.PctUnemployed}{percentage of people 16 and over, in the labor force, and unemployed}
\item{x.PctEmploy}{percentage of people 16 and over who are employed}
\item{x.PctEmplManu}{percentage of people 16 and over who are employed in manufacturing}
\item{x.PctEmplProfServ}{percentage of people 16 and over who are employed in professional services}
\item{x.PctOccupManu}{percentage of people 16 and over who are employed in manufacturing}
\item{x.PctOccupMgmtProf}{percentage of people 16 and over who are employed in management or professional occupations}
\item{x.MalePctDivorce}{percentage of males who are divorced}
\item{x.MalePctNevMarr}{percentage of males who have never married}
\item{x.FemalePctDiv}{percentage of females who are divorced}
\item{x.TotalPctDiv}{percentage of population who are divorced}
\item{x.PersPerFam}{mean number of people per family}
\item{x.PctFam2Par}{percentage of families (with kids) that are headed by two parents}
\item{x.PctKids2Par}{percentage of kids in family housing with two parents}
\item{x.PctYoungKids2Par}{percent of kids 4 and under in two parent households}
\item{x.PctTeen2Par}{percent of kids age 12-17 in two parent households}
\item{x.PctWorkMomYoungKids}{percentage of moms of kids 6 and under in labor force}
\item{x.PctWorkMom}{percentage of moms of kids under 18 in labor force}
\item{x.NumIlleg}{number of kids born to never married}
\item{x.PctIlleg}{percentage of kids born to never married}
\item{x.NumImmig}{total number of people known to be foreign born}
\item{x.PctImmigRecent}{percentage of immigrants who immigated within last 3 years}
\item{x.PctImmigRec5}{percentage of immigrants who immigated within last 5 years}
\item{x.PctImmigRec8}{percentage of immigrants who immigated within last 8 years}
\item{x.PctImmigRec10}{percentage of immigrants who immigated within last 10 years}
\item{x.PctRecentImmig}{percent of population who have immigrated within the last 3 years}
\item{x.PctRecImmig5}{percent of population who have immigrated within the last 5 years}
\item{x.PctRecImmig8}{percent of population who have immigrated within the last 8 years}
\item{x.PctRecImmig10}{percent of population who have immigrated within the last 10 years}
\item{x.PctSpeakEnglOnly}{percent of people who speak only English}
\item{x.PctNotSpeakEnglWell}{percent of people who do not speak English well}
\item{x.PctLargHouseFam}{percent of family households that are large - 6 or more}
\item{x.PctLargHouseOccup}{percent of all occupied households that are large -6 or more people}
\item{x.PersPerOccupHous}{mean persons per household}
\item{x.PersPerOwnOccHous}{mean persons per owner occupied household}
\item{x.PersPerRentOccHous}{mean persons per rental household}
\item{x.PctPersOwnOccup}{percent of people in owner occupied households}
\item{x.PctPersDenseHous}{percent of persons in dense housing -more than 1 person per room}
\item{x.PctHousLess3BR}{percent of housing units with less than 3 bedrooms}
\item{x.MedNumBR}{median number of bedrooms}
\item{x.HousVacant}{number of vacant households}
\item{x.PctHousOccup}{percent of housing occupied}
\item{x.PctHousOwnOcc}{percent of households owner occupied}
\item{x.PctVacantBoarded}{percent of vacant housing that is boarded up}
\item{x.PctVacMore6Mos}{percent of vacant housing that has been vacant more than 6 months}
\item{x.MedYrHousBuilt}{median year housing units built}
\item{x.PctHousNoPhone}{percent of occupied housing units without phone in 1990}
\item{x.PctWOFullPlumb}{percent of housing without complete plumbing facilities}
\item{x.OwnOccLowQuart}{owner occupied housing - lower quartile value}
\item{x.OwnOccMedVal}{owner occupied housing - median value}
\item{x.OwnOccHiQuart}{owner occupied housing - upper quartile value}
\item{x.RentLowQ}{rental housing - lower quartile rent}
\item{x.RentMedian}{rental housing - median rent}
\item{x.RentHighQ}{rental housing - upper quartile rent}
\item{x.MedRent}{median gross rent}
\item{x.MedRentPctHousInc}{median gross rent as a percentage of household income}
\item{x.MedOwnCostPctInc}{median owners cost as a percentage of household income - for owners with a mortgage}
\item{x.MedOwnCostPctIncNoMtg}{median owners cost as a percentage of household income - for owners without a mortgage}
\item{x.NumInShelters}{number of people in homeless shelters}
\item{x.NumStreet}{number of homeless people counted in the street}
\item{x.PctForeignBorn}{percent of people foreign born}
\item{x.PctBornSameState}{percent of people born in the same state as currently living}
\item{x.PctSameHouse85}{percent of people living in the same house as in 1985}
\item{x.PctSameCity85}{percent of people living in the same city as in 1985}
\item{x.PctSameState85}{percent of people living in the same state as in 1985}
\item{x.LemasSwornFT}{number of sworn full time police officers}
\item{x.LemasSwFTPerPop}{sworn full time police officers per 100K population}
\item{x.LemasSwFTFieldOps}{number of sworn full time police officers in field operations}
\item{x.LemasSwFTFieldPerPop}{sworn full time police officers in field operations per 100K population}
\item{x.LemasTotalReq}{total requests for police}
\item{x.LemasTotReqPerPop}{total requests for police per 100K popuation}
\item{x.PolicReqPerOffic}{total requests for police per police officer}
\item{x.PolicPerPop}{police officers per 100K population}
\item{x.RacialMatchCommPol}{a measure of the racial match between the community and the police force. High values indicate proportions in community and police force are similar}
\item{x.PctPolicWhite}{percent of police that are caucasian}
\item{x.PctPolicBlack}{percent of police that are african american}
\item{x.PctPolicHisp}{percent of police that are hispanic}
\item{x.PctPolicAsian}{percent of police that are asian}
\item{x.PctPolicMinor}{percent of police that are minority of any kind}
\item{x.OfficAssgnDrugUnits}{number of officers assigned to special drug units}
\item{x.NumKindsDrugsSeiz}{number of different kinds of drugs seized}
\item{x.PolicAveOTWorked}{police average overtime worked}
\item{x.LandArea}{land area in square miles}
\item{x.PopDens}{population density in persons per square mile}
\item{x.PctUsePubTrans}{percent of people using public transit for commuting}
\item{x.PolicCars}{number of police cars}
\item{x.PolicOperBudg}{police operating budget}
\item{x.LemasPctPolicOnPatr}{percent of sworn full time police officers on patrol}
\item{x.LemasGangUnitDeploy}{gang unit deployed. 0 means NO, 1 means YES, 0.5 means Part Time)}
\item{x.LemasPctOfficDrugUn}{percent of officers assigned to drug units}
\item{x.PolicBudgPerPop}{police operating budget per population}
\item{y.ViolentCrimesPerPop}{total number of violent crimes per 100K popuation  GOAL attribute}
}
}
\source{
{culled from 1990 US Census, 1995 US FBI Uniform Crime Report, 1990 US Law Enforcement Management and Administrative Statistics Survey, available from ICPSR at U of Michigan.}
}
\usage{
Example2
}
\description{
Contains statistics about 123 Communities in 1990.
}
\examples{
data(Example2)
}
\keyword{datasets}
