% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S.DSL2.R
\name{S.DSL2}
\alias{S.DSL2}
\title{The Selected Double Super Learner Method (SDSL): Binary Case Treatment}
\usage{
S.DSL2(
  data,
  y.col.n,
  d.col.n,
  SL.library,
  cv,
  Best.Yt,
  Best.Ytr,
  Best.Dt,
  Best.Dtr
)
}
\arguments{
\item{data}{The data set. Treatment variable D must be binary, and the number of variables must be greater than 3. Variables must be numeric with no missing values.}

\item{y.col.n}{The column number in the data set that represents the response.}

\item{d.col.n}{The column number in the data set that represents the treatment variable.}

\item{SL.library}{The candidate machine-learning algorithms. Must be uploaded first. See example.}

\item{cv}{The number of cross-validation folds followed by the letter L. See example.}

\item{Best.Yt}{The number of the best learner in library in predicting Y using the testing set. For example learner number 1.}

\item{Best.Ytr}{The number of the best learner in library in predicting Y using the training set.}

\item{Best.Dt}{The number of the best learner in library in predicting D using the testing set. For example learner number 5.}

\item{Best.Dtr}{The number of the best learner in library in predicting D using the training set.}
}
\description{
Estimates the treatment effect using the selected DSL method when the treatment variable is considered to be binary.
}
\examples{
DataSet <- DATA2(N=100, p=5, theta=0.5)
SL.library <- c("SL.biglasso", "SL.glm", "SL.kernelKnn", "SL.ranger", "SL.xgboost")
S.DSL2(data=DataSet, y.col.n=1, d.col.n=2, SL.library=SL.library, cv=3L, Best.Yt=1, Best.Ytr=1, Best.Dt=5, Best.Dtr=5)
}
